update: get_youtube_videos TypeScript化
